// laLigaAPI/src/app/interfaces/teams-response.ts
// Generated by https://quicktype.io

export interface TeamsResponse {
    get:        string;
    parameters: Parameters;
    errors:     any[];
    results:    number;
    paging:     Paging;
    response:   Response[];
}

export interface Paging {
    current: number;
    total:   number;
}

export interface Parameters {
    league: string;
    season: string;
}

export interface Response {
    team:  Team;
    venue: Venue;
}

export interface Team {
    id:       number;
    name:     string;
    code:     string;
    country:  Country;
    founded:  number;
    national: boolean;
    logo:     string;
}

export enum Country {
    Spain = "Spain",
}

export interface Venue {
    id:       number;
    name:     string;
    address:  string;
    city:     string;
    capacity: number;
    surface:  Surface;
    image:    string;
}

export enum Surface {
    Grass = "grass",
}